src/backend/requirements.txt
asgiref==3.4.1
black==21.10b0
click==8.0.3
colorama==0.4.4
Django==3.2.8
django-cleanup==5.2.0
django-environ==0.8.1
django-extensions==3.1.5
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pytz==2021.3
regex==2021.11.9
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==3.10.0.2
